Move tracker to screen
